Native bridge updates
